requirements.txt
aiofiles==24.1.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
attrs==24.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
chromedriver==2.24.1
click==8.1.8
cryptography==44.0.2
ecdsa==0.19.0
ezpip==0.2.5
fastapi==0.115.8
fies==1.4.0
fileinit==0.0.2
filelock==3.18.0
fsspec==2025.3.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
jose==1.0.0
json-stock==0.2.6
llvmlite==0.44.0
Mako==1.3.9
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
mult==1.1.1
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai-whisper==20240930
outcome==1.3.0.post0
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
PySocks==1.7.1
pytest==8.3.4
pytest-asyncio==0.25.3
python-jose==3.4.0
python-multipart==0.0.20
regex==2024.11.6
relpath==3.0.5
rename-lock==0.1.4
requests==2.32.3
rsa==4.9
selenium==4.23.1
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sout==1.2.1
SQLAlchemy==2.0.38
starlette==0.45.3
sympy==1.13.1
tiktoken==0.9.0
timedelta==2020.12.3
torch==2.6.0
tqdm==4.67.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.34.0
webdrivermanager-cn==1.8.5
websocket-client==1.8.0
wsproto==1.2.0
